ML/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.9.2
langchain==0.3.7
langchain-openai==0.2.6
pypdf==5.1.0
python-docx==1.1.2
pdfminer.six==20231228
httpx==0.27.2
langchain-google-genai==2.0.4
google-generativeai==0.8.3

